requirements.txt
pyswisseph>=2.10.0
pydantic>=2.0.0
